refactor gadget_finder
