+++ tests/test_find_gadgets.py
@@ -18,20 +18,18 @@ If the gadget is completely missing after your changes. Pick the address that di
 The logging should say why the gadget was discarded.
 
 rop = p.analyses.ROP()
-angrop.gadget_analyzer.l.setLevel("DEBUG")
-rop._gadget_analyzer.analyze_gadget(addr)
+rop.analyze_gadget(addr)
 
 If a gadget is missing memory reads / memory writes / memory changes, the actions are probably missing.
 Memory changes require a read action followed by a write action to the same address.
 """
 
 def gadget_exists(rop, addr):
-    return rop._gadget_analyzer.analyze_gadget(addr) is not None
+    return rop.analyze_gadget(addr) is not None
 
 def test_badbyte():
     proj = angr.Project(os.path.join(tests_dir, "i386", "bronze_ropchain"), auto_load_libs=False)
     rop = proj.analyses.ROP()
-    rop._initialize_gadget_analyzer()
 
     assert all(gadget_exists(rop, x) for x in [0x080a9773, 0x08091cf5, 0x08092d80, 0x080920d3])
 
@@ -46,16 +44,14 @@ def local_multiprocess_find_gadgets():
 def test_symbolic_memory_access_from_stack():
     proj = angr.Project(os.path.join(tests_dir, "armel", "test_angrop_arm_gadget"), auto_load_libs=False)
     rop = proj.analyses.ROP()
-    rop._initialize_gadget_analyzer()
 
     assert all(gadget_exists(rop, x) for x in [0x000103f4])
 
 def test_arm_thumb_mode():
     proj = angr.Project(os.path.join(bin_path, "tests", "armel", "libc-2.31.so"), auto_load_libs=False)
     rop = proj.analyses.ROP(fast_mode=False, only_check_near_rets=False, is_thumb=True)
-    rop._initialize_gadget_analyzer()
 
-    gadget = rop._gadget_analyzer.analyze_gadget(0x4bf858+1)
+    gadget = rop.analyze_gadget(0x4bf858+1)
 
     assert gadget
     assert gadget.block_length == 6
